requirements.txt
# LangChain 核心
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-community>=0.4.0
langgraph>=1.0.0
langchain-text-splitters
# Web 框架
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9

# 前端界面 (可选，保留原 Gradio 版本)
gradio>=4.0.0

# 向量数据库
chromadb>=0.4.0

# 文档解析
pymupdf>=1.23.0
python-docx>=1.0.0

# Embedding 模型
sentence-transformers>=2.2.0

# 其他工具
python-dotenv>=1.0.0
tiktoken>=0.5.0
